refactor: endpoint반영
